requirements.txt
# --------- pytorch --------- #
torchvision
lightning
omegaconf
torchmetrics[image]
#wandb-osh==1.2.1
peft
timm
openai-clip
pytorchvideo
einops>=0.8.0
rotary_embedding_torch
diffusers
rotary-embedding-torch
transformers
tfrecord[torch]
gluonts[torch]==0.13.1
colorama
tqdm
opencv-python
matplotlib
click
moviepy
imageio
pandas
pyzmq
pyrealsense2
internetarchive
h5py
pyyaml
scipy
bezier
zmq
pyiqa==0.1.10
easydict
pytubefix
roma

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb[media]
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)